pyproject.toml
[tool.poetry]
package-mode = false
name = "hackathon-bot"
version = "0.1.0"
description = ""
authors = ["iamnovichek <rungotika@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.1.2"
aiogram = "^3.13.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
black = "^24.10.0"
isort = "^5.13.2"
requests = "^2.32.3"
python-dateutil = "^2.9.0.post0"
gunicorn = "^23.0.0"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
redis = "^5.2.0"
faker = "^33.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
